+++ Source/GSWindowDecorationView.m
@@ -80,13 +80,13 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
 
 + (void) offsets: (float *)l : (float *)r : (float *)t : (float *)b
-    forStyleMask: (unsigned int)style
+    forStyleMask: (NSUInteger)style
 {
   [self subclassResponsibility: _cmd];
 }
 
 + (NSRect) contentRectForFrameRect: (NSRect)aRect
-			 styleMask: (unsigned int)aStyle
+			 styleMask: (NSUInteger)aStyle
 {
   float t = 0.0, b = 0.0, l = 0.0, r = 0.0;
 
@@ -108,7 +108,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 }
 
 + (NSRect) frameRectForContentRect: (NSRect)aRect
-			 styleMask: (unsigned int)aStyle
+			 styleMask: (NSUInteger)aStyle
 {
   float t = 0.0, b = 0.0, l = 0.0, r = 0.0;
 
@@ -130,8 +130,8 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
   return aRect;
 }
 
-+ (float) minFrameWidthWithTitle: (NSString *)aTitle
-		       styleMask: (unsigned int)aStyle
++ (CGFloat) minFrameWidthWithTitle: (NSString *)aTitle
+                         styleMask: (NSUInteger)aStyle
 {
   [self subclassResponsibility: _cmd];
   return 0.0;
@@ -183,7 +183,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 }
 
 - (NSRect) contentRectForFrameRect: (NSRect)aRect
-                         styleMask: (unsigned int)aStyle
+                         styleMask: (NSUInteger)aStyle
 {
   NSRect content = [object_getClass(self) 
                        contentRectForFrameRect: aRect
@@ -192,9 +192,8 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
   if ([_window menu] != nil)
     {
-      float	menubarHeight = [[GSTheme theme] 
-					 menuHeightForWindow: 
-				    _window];
+      CGFloat menubarHeight = 
+        [[GSTheme theme] menuHeightForWindow: _window];
   
       content.size.height -= menubarHeight;
     }
@@ -210,15 +209,14 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 }
 
 - (NSRect) frameRectForContentRect: (NSRect)aRect
-                         styleMask: (unsigned int)aStyle
+                         styleMask: (NSUInteger)aStyle
 {
   NSToolbar *tb = [_window toolbar];
 
   if ([_window menu] != nil)
     {
-      float	menubarHeight = [[GSTheme theme] 
-				  menuHeightForWindow: 
-				    _window];
+      CGFloat menubarHeight =
+        [[GSTheme theme] menuHeightForWindow: _window];
 
       aRect.size.height += menubarHeight;
     }
@@ -296,7 +294,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
     {
       NSMenuView *menuView = nil;
       GSTheme *theme = [GSTheme theme];
-      float menuBarHeight = [theme menuHeightForWindow: _window];
+      CGFloat menuBarHeight = [theme menuHeightForWindow: _window];
       NSRect menuRect = 
 	NSMakeRect(contentViewFrame.origin.x,
 		   (NSMaxY(contentRect) + 1) - menuBarHeight,  
@@ -320,7 +318,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
   if (hasToolbar)
     {
       GSToolbarView *tv = [tb _toolbarView];
-      float newToolbarViewHeight;
+      CGFloat newToolbarViewHeight;
       NSRect toolbarRect;
       // If the width changed we may need to recalculate the height
       if (contentViewFrame.size.width != [tv frame].size.width)
@@ -351,7 +349,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
     }
 }
 
-- (void) changeWindowHeight: (float)difference
+- (void) changeWindowHeight: (CGFloat)difference
 {
   NSRect orgWindowFrame;
   NSRect windowFrame;
@@ -451,8 +449,8 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
 - (void) addToolbarView: (GSToolbarView*)toolbarView
 {
-  float newToolbarViewHeight;
-  float contentYOrigin;
+  CGFloat newToolbarViewHeight;
+  CGFloat contentYOrigin;
 
   hasToolbar = YES;
   [toolbarView setFrameSize: NSMakeSize(contentRect.size.width, 100)];
@@ -464,7 +462,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
   contentYOrigin = NSMaxY(contentRect);
   if (hasMenu)
     {
-      float menuBarHeight = [[GSTheme theme] menuHeightForWindow: _window];
+      CGFloat menuBarHeight = [[GSTheme theme] menuHeightForWindow: _window];
       contentYOrigin -= menuBarHeight;
     }
 
@@ -481,7 +479,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
 - (void) removeToolbarView: (GSToolbarView *)toolbarView
 {
-  float toolbarViewHeight = [toolbarView frame].size.height;
+  CGFloat toolbarViewHeight = [toolbarView frame].size.height;
 
   // Unplug the toolbar view
   hasToolbar = NO;
@@ -494,8 +492,8 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 {
   // Frame and height
   NSRect toolbarViewFrame = [toolbarView frame];
-  float toolbarViewHeight = toolbarViewFrame.size.height;
-  float newToolbarViewHeight = [toolbarView _heightFromLayout];
+  CGFloat toolbarViewHeight = toolbarViewFrame.size.height;
+  CGFloat newToolbarViewHeight = [toolbarView _heightFromLayout];
   
   if (toolbarViewHeight != newToolbarViewHeight)
     {
@@ -515,7 +513,7 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
 - (void) addMenuView: (NSMenuView*)menuView
 {
-  float	menubarHeight = [[GSTheme theme] 
+  CGFloat menubarHeight = [[GSTheme theme] 
 			  menuHeightForWindow: 
 			    _window];
   NSRect menuRect = NSMakeRect(contentRect.origin.x,
@@ -532,9 +530,9 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 
 - (NSMenuView*) removeMenuView
 {
-  NSEnumerator	*e = [[self subviews] objectEnumerator];
-  NSView	*v;
-  float	menubarHeight = [[GSTheme theme] 
+  NSEnumerator *e = [[self subviews] objectEnumerator];
+  NSView *v;
+  CGFloat menubarHeight = [[GSTheme theme] 
 			  menuHeightForWindow: 
 			    _window];
   
@@ -563,13 +561,13 @@ static inline NSRect RectWithSizeScaledByFactor(NSRect aRect, CGFloat factor)
 @implementation GSBackendWindowDecorationView
 
 + (void) offsets: (float *)l : (float *)r : (float *)t : (float *)b
-    forStyleMask: (unsigned int)style
+    forStyleMask: (NSUInteger)style
 {
   [GSCurrentServer() styleoffsets: l : r : t : b : style];
 }
 
-+ (float) minFrameWidthWithTitle: (NSString *)aTitle
-		       styleMask: (unsigned int)aStyle
++ (CGFloat) minFrameWidthWithTitle: (NSString *)aTitle
+                         styleMask: (NSUInteger)aStyle
 {
   /* TODO: we could at least guess... */
   return 0.0;
