+++ Source/NSPrintPanel.m
@@ -124,7 +124,7 @@ static NSPrintPanel *shared_instance = nil;
 
 /* Designated initializer */
 - (id) initWithContentRect: (NSRect)contentRect
-		 styleMask: (unsigned int)aStyle
+		 styleMask: (NSUInteger)aStyle
 		   backing: (NSBackingStoreType)bufferingType
 		     defer: (BOOL)flag
 {
